changes
Former-commit-id: 8fed1551291aa244425ca0b2089c631e5f28a828

+++ setup.py
@@ -10,13 +10,13 @@ from packaging.version import LegacyVersion
 from skbuild.exceptions import SKBuildError
 from skbuild.cmaker import get_cmake_version
 
-if not os.path.exists(os.path.join("src", "librapid", "pybind11")):
+if os.path.exists("pybind11") and not os.path.exists(os.path.join("src", "librapid", "pybind11")):
 	shutil.copytree("pybind11", os.path.join("src", "librapid", "pybind11"))
 
-if not os.path.exists(os.path.join("src", "librapid", "jitify")):
+if os.path.exists("jitify") and not os.path.exists(os.path.join("src", "librapid", "jitify")):
 	shutil.copytree("jitify", os.path.join("src", "librapid", "jitify"))
 
-if not os.path.exists(os.path.join("src", "librapid", "openblas_install")):
+if os.path.exists("openblas_install") and not os.path.exists(os.path.join("src", "librapid", "openblas_install")):
 	shutil.copytree("openblas_install", os.path.join("src", "librapid", "openblas_install"))
 
 # if os.path.exists("_skbuild"):
